scrutinizer issues fixes

+++ src/Adapter.php
@@ -21,6 +21,11 @@ class Adapter extends AbstractAdapter
     protected $bucket;
 
     /**
+     * @var
+     */
+    protected $debug;
+
+    /**
      * Adapter constructor.
      *
      * @param $config
@@ -32,6 +37,7 @@ class Adapter extends AbstractAdapter
         Conf::setSecretKey($config['secret_key']);
 
         $this->bucket = $config['bucket'];
+        $this->debug = $config['debug'];
 
         $this->setPathPrefix($config['protocol'].'://'.$config['domain'].'/');
 
@@ -39,7 +45,7 @@ class Adapter extends AbstractAdapter
     }
 
     /**
-     * @return mixed
+     * @return string
      */
     public function getBucket()
     {
@@ -47,7 +53,7 @@ class Adapter extends AbstractAdapter
     }
 
     /**
-     * @param $path
+     * @param string $path
      *
      * @return string
      */
@@ -63,26 +69,31 @@ class Adapter extends AbstractAdapter
      *
      * @throws RuntimeException
      *
-     * @return bool
+     * @return array|bool
      */
     public function write($path, $contents, Config $config)
     {
         $tmpfname = $this->writeTempFile($contents);
 
+        if (false === $tmpfname) {
+            return false;
+        }
+
         try {
-            $response = $this->normalizeResponse(
-                Cosapi::upload($this->getBucket(), $tmpfname, $path)
-            );
+            $response = Cosapi::upload($this->getBucket(), $tmpfname, $path,
+                                        null, null, $config->get('insertOnly', 1));
 
             $this->deleteTempFile($tmpfname);
+
+            $response = $this->normalizeResponse($response);
+
+            if (false === $response) {
+                return false;
+            }
 
             $this->setContentType($path, $contents);
         } catch (RuntimeException $exception) {
             $this->deleteTempFile($tmpfname);
-
-            if ($exception->getCode() == -4018) {
-                return $this->getMetadata($path);
-            }
 
             throw $exception;
         }
@@ -95,25 +106,24 @@ class Adapter extends AbstractAdapter
      * @param resource $resource
      * @param Config   $config
      *
-     * @return bool
+     * @throws RuntimeException
+     *
+     * @return array|bool
      */
     public function writeStream($path, $resource, Config $config)
     {
         $uri = stream_get_meta_data($resource)['uri'];
 
-        try {
-            $response = $this->normalizeResponse(
-                Cosapi::upload($this->getBucket(), $uri, $path)
-            );
+        $response = Cosapi::upload($this->getBucket(), $uri, $path,
+                                    null, null, $config->get('insertOnly', 1));
 
-            $this->setContentType($path, stream_get_contents($resource));
-        } catch (RuntimeException $exception) {
-            if ($exception->getCode() == -4018) {
-                return $this->getMetadata($path);
-            }
+        $response = $this->normalizeResponse($response);
 
-            throw $exception;
+        if (false === $response) {
+            return false;
         }
+
+        $this->setContentType($path, stream_get_contents($resource));
 
         return $response;
     }
@@ -123,26 +133,33 @@ class Adapter extends AbstractAdapter
      * @param string $contents
      * @param Config $config
      *
-     * @return bool
+     * @throws RuntimeException
+     *
+     * @return array|bool
      */
     public function update($path, $contents, Config $config)
     {
         $tmpfname = $this->writeTempFile($contents);
 
+        if (false === $tmpfname) {
+            return false;
+        }
+
         try {
-            $response = $this->normalizeResponse(
-                Cosapi::upload($this->getBucket(), $tmpfname, $path, null, null, 0)
-            );
+            $response = Cosapi::upload($this->getBucket(), $tmpfname, $path,
+                                        null, null, $config->get('insertOnly', 0));
 
             $this->deleteTempFile($tmpfname);
+
+            $response = $this->normalizeResponse($response);
+
+            if (false === $response) {
+                return false;
+            }
 
             $this->setContentType($path, $contents);
         } catch (RuntimeException $exception) {
             $this->deleteTempFile($tmpfname);
-
-            if ($exception->getCode() == -4018) {
-                return $this->getMetadata($path);
-            }
 
             throw $exception;
         }
@@ -155,25 +172,24 @@ class Adapter extends AbstractAdapter
      * @param resource $resource
      * @param Config   $config
      *
-     * @return bool
+     * @throws RuntimeException
+     *
+     * @return array|bool
      */
     public function updateStream($path, $resource, Config $config)
     {
         $uri = stream_get_meta_data($resource)['uri'];
 
-        try {
-            $response = $this->normalizeResponse(
-                Cosapi::upload($this->getBucket(), $uri, $path, $path, null, null, 0)
-            );
+        $response = Cosapi::upload($this->getBucket(), $uri, $path,
+                                    null, null, $config->get('insertOnly', 0));
 
-            $this->setContentType($path, stream_get_contents($resource));
-        } catch (RuntimeException $exception) {
-            if ($exception->getCode() == -4018) {
-                return $this->getMetadata($path);
-            }
+        $response = $this->normalizeResponse($response);
 
-            throw $exception;
+        if (false === $response) {
+            return false;
         }
+
+        $this->setContentType($path, stream_get_contents($resource));
 
         return $response;
     }
@@ -186,7 +202,7 @@ class Adapter extends AbstractAdapter
      */
     public function rename($path, $newpath)
     {
-        return $this->normalizeResponse(
+        return (bool) $this->normalizeResponse(
             Cosapi::move($this->getBucket(), $path, $newpath, 1)
         );
     }
@@ -201,7 +217,11 @@ class Adapter extends AbstractAdapter
     {
         $resource = $this->read($path);
 
-        return $this->update($newpath, $resource['contents'], new Config());
+        if (false === $resource) {
+            return false;
+        }
+
+        return (bool) $this->update($newpath, $resource['contents'], new Config());
     }
 
     /**
@@ -211,7 +231,7 @@ class Adapter extends AbstractAdapter
      */
     public function delete($path)
     {
-        return $this->normalizeResponse(
+        return (bool) $this->normalizeResponse(
             Cosapi::delFile($this->getBucket(), $path)
         );
     }
@@ -223,7 +243,7 @@ class Adapter extends AbstractAdapter
      */
     public function deleteDir($dirname)
     {
-        return $this->normalizeResponse(
+        return (bool) $this->normalizeResponse(
             Cosapi::delFolder($this->getBucket(), $dirname)
         );
     }
@@ -232,7 +252,7 @@ class Adapter extends AbstractAdapter
      * @param string $dirname
      * @param Config $config
      *
-     * @return bool
+     * @return array|bool
      */
     public function createDir($dirname, Config $config)
     {
@@ -245,13 +265,13 @@ class Adapter extends AbstractAdapter
      * @param string $path
      * @param string $visibility
      *
-     * @return mixed
+     * @return bool
      */
     public function setVisibility($path, $visibility)
     {
         $visibility = $visibility === AdapterInterface::VISIBILITY_PUBLIC ? 'eWPrivateRPublic' : 'eWRPrivate';
 
-        return $this->normalizeResponse(
+        return (bool) $this->normalizeResponse(
             Cosapi::update($this->getBucket(), $path, null, $visibility)
         );
     }
@@ -264,18 +284,16 @@ class Adapter extends AbstractAdapter
     public function has($path)
     {
         try {
-            $this->getMetadata($path);
+            return (bool) $this->getMetadata($path);
         } catch (RuntimeException $exception) {
             return false;
         }
-
-        return true;
     }
 
     /**
      * @param string $path
      *
-     * @return array
+     * @return string
      */
     public function read($path)
     {
@@ -320,71 +338,73 @@ class Adapter extends AbstractAdapter
     /**
      * @param string $path
      *
-     * @return array
+     * @return array|bool
      */
     public function getSize($path)
     {
         $stat = $this->getMetadata($path);
 
-        if ($stat) {
+        if (isset($stat['filesize'])) {
             return ['size' => $stat['filesize']];
         }
 
-        return ['size' => 0];
+        return false;
     }
 
     /**
      * @param string $path
      *
-     * @return array
+     * @return array|bool
      */
     public function getMimetype($path)
     {
         $stat = $this->getMetadata($path);
 
-        if ($stat && !empty($stat['custom_headers']) && !empty($stat['custom_headers']['Content-Type'])) {
+        if (isset($stat['custom_headers']['Content-Type'])) {
             return ['mimetype' => $stat['custom_headers']['Content-Type']];
         }
 
-        return ['mimetype' => ''];
+        return false;
     }
 
     /**
      * @param string $path
      *
-     * @return array
+     * @return array|bool
      */
     public function getTimestamp($path)
     {
         $stat = $this->getMetadata($path);
 
-        if ($stat) {
+        if (isset($stat['ctime'])) {
             return ['timestamp' => $stat['ctime']];
         }
 
-        return ['timestamp' => 0];
+        return false;
     }
 
     /**
      * @param string $path
      *
-     * @return array
+     * @return array|bool
      */
     public function getVisibility($path)
     {
         $stat = $this->getMetadata($path);
 
-        $visibility = AdapterInterface::VISIBILITY_PRIVATE;
-
-        if ($stat && $stat['authority'] === 'eWPrivateRPublic') {
-            $visibility = AdapterInterface::VISIBILITY_PUBLIC;
+        if (isset($stat['authority']) && $stat['authority'] === 'eWPrivateRPublic') {
+            return ['visibility' => AdapterInterface::VISIBILITY_PUBLIC];
         }
 
-        return ['visibility' => $visibility];
+        if (isset($stat['authority']) && $stat['authority'] === 'eWRPrivate') {
+            return ['visibility' => AdapterInterface::VISIBILITY_PRIVATE];
+        }
+
+        return false;
     }
 
     /**
-     * @param $content
+     * @param string $content
      *
      * @return bool|string
      */
@@ -400,16 +420,22 @@ class Adapter extends AbstractAdapter
     }
 
     /**
-     * @param $tmpfname
+     * @param string|bool $tmpfname
+     *
+     * @return bool
      */
     private function deleteTempFile($tmpfname)
     {
+        if (false === $tmpfname) {
+            return false;
+        }
+
         return unlink($tmpfname);
     }
 
     /**
-     * @param $path
-     * @param $content
+     * @param string $path
+     * @param string $content
      *
      * @return bool
      */
@@ -433,12 +459,14 @@ class Adapter extends AbstractAdapter
      */
     protected function normalizeResponse($response)
     {
-        $response = is_array($response) ? $response : json_decode($response, true);
-
         if ($response['code'] == 0) {
             return isset($response['data']) ? $response['data'] : true;
         }
 
-        throw new RuntimeException($response['message'], $response['code']);
+        if ($this->debug) {
+            throw new RuntimeException($response['message'], $response['code']);
+        }
+
+        return false;
     }
 }
